added tests to addItem, failing
